removed unrelated files and moved everything out of the VSCodeLua folder into a FABRIK folder. More progres made on the solver, the angle and joint restrictions in the backwards step are still wrong
